Pushing refactor of the actions, but have now broken tests
