+++ frontend/src/User/Login.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { StyleSheet, css } from 'aphrodite';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { requestLogin, clearLoginError } from './loginActions';
+import { requestLogin, clearLoginError } from './userActions';
 
 class Login extends Component {
   static propTypes = {
